+++ buildSrc/src/main/kotlin/dependency/script/AndroidTools.kt
@@ -6,5 +6,5 @@ object AndroidTools : Dependency {
 
     override val group = "com.android.tools.build"
     override val artifact = "gradle"
-    override val version = "3.5.0"
+    override val version = "3.5.2"
 }