+++ buildSrc/src/main/kotlin/dependency/script/AtomicFu.kt
@@ -7,5 +7,5 @@ object AtomicFu: Dependency {
 
     override val group = "org.jetbrains.kotlinx"
     override val artifact = "atomicfu"
-    override val version = "0.13.0"
+    override val version = "0.14.1"
 }