+++ core/src/commonMain/kotlin/com/algolia/instantsearch/core/hits/HitsConnection.kt
@@ -1,12 +1,13 @@
 package com.algolia.instantsearch.core.hits
 
+import com.algolia.instantsearch.core.Presenter
 import com.algolia.instantsearch.core.connection.Connection
 import com.algolia.instantsearch.core.searcher.Searcher
 
 
 public fun <R, T> Searcher<R>.connectHitsView(
     adapter: HitsView<T>,
-    transform: (R) -> List<T>
+    presenter: Presenter<R, List<T>>
 ): Connection {
-    return HitsConnectionView(this, adapter, transform)
+    return HitsConnectionView(this, adapter, presenter)
 }