+++ helper/src/androidTest/kotlin/stats/TestStatsTextView.kt
@@ -1,5 +1,6 @@
 package stats
 
+import android.os.Build
 import android.widget.TextView
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.SmallTest
@@ -12,6 +13,7 @@ import shouldEqual
 
 
 @SmallTest
+@Config(sdk = [Build.VERSION_CODES.P])
 @RunWith(AndroidJUnit4::class)
 class TestStatsTextView {