+++ helper/src/androidTest/kotlin/stats/TestStatsTextViewSpanned.kt
@@ -1,5 +1,6 @@
 package stats
 
+import android.os.Build
 import android.text.SpannedString
 import android.widget.TextView
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -8,10 +9,12 @@ import applicationContext
 import com.algolia.instantsearch.helper.android.stats.StatsTextViewSpanned
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.robolectric.annotation.Config
 import shouldEqual
 
 
 @SmallTest
+@Config(sdk = [Build.VERSION_CODES.P])
 @RunWith(AndroidJUnit4::class)
 class TestStatsTextViewSpanned {